Expose MapDefinitionRepository
