+++ src/main/java/io/luna/game/model/def/DefinitionRepository.java
@@ -5,8 +5,6 @@ import com.google.common.collect.UnmodifiableIterator;
 
 import java.util.Collection;
 import java.util.Iterator;
-import java.util.LinkedHashMap;
-import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Objects;
 import java.util.Optional;
@@ -26,32 +24,6 @@ import static com.google.common.base.Preconditions.checkState;
  * @author lare96 <http://github.com/lare96>
  */
 public abstract class DefinitionRepository<T extends Definition> implements Iterable<T> {
-
-    /**
-     * An implementation of {@link DefinitionRepository} that is backed internally by a {@link LinkedHashMap}.
-     */
-    static final class MapDefinitionRepository<V extends Definition> extends DefinitionRepository<V> {
-
-        /**
-         * A map of definitions.
-         */
-        private final Map<Integer, V> definitions = new LinkedHashMap<>();
-
-        @Override
-        void put(int id, V definition) {
-            definitions.put(id, definition);
-        }
-
-        @Override
-        public Optional<V> get(int id) {
-            return Optional.ofNullable(definitions.get(id));
-        }
-
-        @Override
-        public Iterator<V> newIterator() {
-            return definitions.values().iterator();
-        }
-    }
 
     /**
      * If this repository is read-only.
